+++ react/__mocks__/vtex.store-components/Container.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react'
+
+/**
+ * Container Mocked Component.
+ */
+export default class Container extends Component {
+  render() {
+    return (
+      <div>Container</div>
+    )
+  }
+}